requirements.txt
django==1.6.5
anyjson==0.3.3
ipython==2.1.0
wsgiref==0.1.2
markdown==2.4.1
djangorestframework==2.3.14
django-filter==0.7
